backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.32.3

# Phase 1 dependencies
numpy==1.24.3
chromadb==0.4.15

# Testing
pytest==7.4.3


